+++ Videogame/QuetzalUnity/Assets/Prefabs/EndgameManager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 08a132a3d8b5f41ceb8ea4c2f13b8aad
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
