+++ Videogame/QuetzalUnity/Assets/Scripts/Boss/BOSSHealth.cs
@@ -5,8 +5,6 @@ using UnityEngine.SceneManagement;
 
 public class BOSSHealth : MonoBehaviour
 {
-    public Canvas credits;
-    public Canvas canvas;
     public bool isTransformed = false;
     public float enemyHealth;
     public float maxHealth;
@@ -17,9 +15,6 @@ public class BOSSHealth : MonoBehaviour
     // Start is called before the first frame update
     void Start()
     {   
-        credits = GameObject.Find("CanvasCredits").GetComponent<Canvas>();
-        canvas = GameObject.Find("Canvas").GetComponent<Canvas>();
-        credits.enabled = false;
         maxHealth = GetComponent<BossStats>().enemyHealth;
         enemyHealth = maxHealth;
         gameManager = GameObject.Find("EndgameManager");
@@ -39,15 +34,8 @@ public class BOSSHealth : MonoBehaviour
         if (enemyHealth <= 0)
         {
             gameManager.GetComponent<EndGame>().bossIsDead = true;
-            canvas.enabled = false;
-            credits.enabled = true;
+            gameManager.GetComponent<bossManager>().bossIsDead = true;
+            Destroy(gameObject);
         }        
-    }
-
-    IEnumerator Wait()
-    {
-        yield return new WaitForSeconds(20);
-        SceneManager.LoadScene(2, LoadSceneMode.Single);
-        Destroy(gameObject);
     }
 }
