+++ Videogame/QuetzalUnity/Assets/Scripts/bossManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5b707626b422c493b9dae77407567f0d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
