+++ Videogame/QuetzalUnity/UserSettings/EditorUserSettings.asset
@@ -15,13 +15,13 @@ EditorUserSettings:
       value: 515250075c0c595e5f5a5e71122159444e4e4a2f7a7d7f602f284d66b4b76661
       flags: 0
     RecentlyUsedSceneGuid-3:
-      value: 0709555203570a02595d587148770a40154e1e7d7c292431787d4b62b3b4306c
-      flags: 0
-    RecentlyUsedSceneGuid-4:
       value: 5b010c5656565959095d0e2614775b14144f1e7c2d7020697e7f4962b3b4326a
       flags: 0
+    RecentlyUsedSceneGuid-4:
+      value: 0709555203570a02595d587148770a40154e1e7d7c292431787d4b62b3b4306c
+      flags: 0
     RecentlyUsedSceneGuid-5:
-      value: 550806520754590b5e5e08204377074044164b2b752b75352e791f32e4b9326f
+      value: 0709555203570a02595d587148770a40154e1e7d7c292431787d4b62b3b4306c
       flags: 0
     RecentlyUsedSceneGuid-6:
       value: 550806520754590b5e5e08204377074044164b2b752b75352e791f32e4b9326f
